Lab4/Bot_Test/requirements.txt
python-telegram-bot>=20.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
pytz>=2023.3

